Rename Client to RDSProvisioningClient
